+++ lib/src/adaptyui.dart
@@ -1,4 +1,5 @@
 import 'dart:convert' show json;
+import 'package:adapty_ui_flutter/src/models/adaptyui_action.dart';
 import 'package:adapty_ui_flutter/src/models/adaptyui_view.dart';
 import 'package:flutter/services.dart';
 import 'package:adapty_flutter/adapty_flutter.dart';
@@ -27,7 +28,7 @@ class AdaptyUI {
 
   AdaptyUI._internal();
 
-  static const String sdkVersion = '1.1.0';
+  static const String sdkVersion = '2.0.0';
 
   static const String _channelName = 'flutter.adapty.com/adapty_ui';
   static const MethodChannel _channel = MethodChannel(_channelName);
@@ -53,26 +54,13 @@ class AdaptyUI {
   /// - an [AdaptyUIView] object, representing the requested paywall screen.
   Future<AdaptyUIView> createPaywallView({
     required AdaptyPaywall paywall,
+    required String locale,
     bool preloadProducts = false,
-    AdaptyUIProductsTitlesResolver? productsTitlesResolver,
   }) async {
-    Map<String, String>? productsTitles;
-
-    if (productsTitlesResolver != null) {
-      productsTitles = <String, String>{};
-
-      for (var productId in paywall.vendorProductIds) {
-        final title = productsTitlesResolver(productId);
-        if (title != null) productsTitles[productId] = title;
-      }
-    } else {
-      productsTitles = null;
-    }
-
     final result = (await _invokeMethodHandlingErrors<String>(Method.createView, {
       Argument.paywall: json.encode(paywall.jsonValue),
+      Argument.locale: locale,
       Argument.preloadProducts: preloadProducts,
-      if (productsTitles != null) Argument.productsTitles: productsTitles,
     })) as String;
 
     final view = AdaptyUIViewJSONBuilder.fromJsonValue(json.decode(result));
@@ -135,11 +123,15 @@ class AdaptyUI {
     final view = AdaptyUIViewJSONBuilder.fromJsonValue(json.decode(call.arguments[Argument.view]));
 
     switch (call.method) {
-      case Method.paywallViewDidPressCloseButton:
-        _observer!.paywallViewDidPressCloseButton(view);
+      case Method.paywallViewDidPerformAction:
+        final action = AdaptyUIActionJSONBuilder.fromJsonValue(json.decode(call.arguments[Argument.action]));
+        _observer!.paywallViewDidPerformAction(view, action);
         break;
       case Method.paywallViewDidPerformSystemBackAction:
-        _observer!.paywallViewDidPerformSystemBackActionOnAndroid(view);
+        _observer!.paywallViewDidPerformAction(
+          view,
+          const AdaptyUIAction(AdaptyUIActionType.androidSystemBack, null),
+        );
         break;
       case Method.paywallViewDidSelectProduct:
         final product = AdaptyPaywallProductJSONBuilder.fromJsonValue(json.decode(call.arguments[Argument.product]));
@@ -176,25 +168,8 @@ class AdaptyUI {
         _observer!.paywallViewDidFailRendering(view, error);
         break;
       case Method.paywallViewDidFailLoadingProducts:
-        AdaptyIOSProductsFetchPolicy? fetchPolicy;
-
-        if (AdaptySDKNative.isIOS) {
-          final String fetchPolicyString = call.arguments[Argument.fetchPolicy];
-
-          switch (fetchPolicyString) {
-            case 'wait_for_receipt_validation': // TODO: use Adapty-SDK
-              fetchPolicy = AdaptyIOSProductsFetchPolicy.waitForReceiptValidation;
-              break;
-            default:
-              fetchPolicy = AdaptyIOSProductsFetchPolicy.defaultPolicy;
-              break;
-          }
-        } else {
-          fetchPolicy = null;
-        }
-
         final error = AdaptyErrorJSONBuilder.fromJsonValue(json.decode(call.arguments[Argument.error]));
-        _observer!.paywallViewDidFailLoadingProducts(view, fetchPolicy, error);
+        _observer!.paywallViewDidFailLoadingProducts(view, error);
         break;
       default:
         break;