+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.0.0
+
+🎉 We are happy to introduce our new version of AdaptyUI SDK! Please, revise our [documentation](https://docs.adapty.io/docs/paywall-builder-installation-flutter).
+
+⚠️ **Breaking Changes**:
+
+- `locale` param introduced for `.createPaywallView()` method.
+- `productsTitlesResolver` param of `.createPaywallView()` method was removed. This feature was implemented within dashboard.
+- Introducing `AdaptyUIAction`. [Read more.](https://docs.adapty.io/docs/paywall-builder-events-flutter#actions)
+- `.paywallViewDidPressCloseButton()` was replaced with `.paywallViewDidPerformAction()`.
+- `.paywallViewDidPerformSystemBackActionOnAndroid` was replaced with `.paywallViewDidPerformAction()`.
+
 ## 1.1.0
 
 - Initial Release. [Read More.](https://docs.adapty.io/docs/paywall-builder-getting-started)